Fix circularity caused by getTags
